Added Save button
Added functionality to save the active paint slot as an image to disk. Will call .save_dirty() if the image has already been previously saved to disk.

Known Issue: if you delete the saved image while the .blend is loaded, the save acts like it works, but does nothing. Not sure if it's my code or a Blender thing.